Source snapshot from Powershell/openssh-portable:latestw_all

+++ contrib/win32/openssh/OpenSSHBuildHelper.psm1
@@ -300,8 +300,8 @@ function Package-OpenSSH
     }
     $buildDir = Join-Path $repositoryRoot ("bin\" + $folderName + "\" + $Configuration)
     $payload = "sshd.exe", "ssh.exe", "ssh-agent.exe", "ssh-add.exe", "sftp.exe"
-    $payload += "sftp-server.exe", "scp.exe", "ssh-shellhost.exe", "ssh-keygen.exe" 
-    $payload += "sshd_config", "install-sshd.ps1", "uninstall-sshd.ps1"
+    $payload += "sftp-server.exe", "scp.exe", "ssh-shellhost.exe", "ssh-keygen.exe", "ssh-keyscan.exe" 
+    $payload += "sshd_config", "install-sshd.ps1", "uninstall-sshd.ps1", "FixHostFilePermissions.ps1", "FixUserFilePermissions.ps1", "OpenSSHUtils.psm1"
 
     $packageName = "OpenSSH-Win64"
     if ($NativeHostArch -eq 'x86') {
@@ -339,7 +339,7 @@ function Package-OpenSSH
 
     if ($DestinationPath -ne "") {
         if (Test-Path $DestinationPath) {            
-            Remove-Item $DestinationPath\* -Force -Recurse
+            Remove-Item $DestinationPath\* -Force -Recurse -ErrorAction SilentlyContinue
         }
         else {
             New-Item -ItemType Directory $DestinationPath -Force | Out-Null
@@ -503,7 +503,7 @@ function Install-OpenSSH
     & "$OpenSSHDir\ssh-keygen.exe" -A
 
     $keyFiles = Get-ChildItem "$OpenSSHDir\ssh_host_*_key*" | % {        
-        Add-PermissionToFileACL -FilePath $_.FullName -User "NT Service\sshd" -Perm "Read"
+        Adjust-HostKeyFileACL -FilePath $_.FullName
     }
 
 
@@ -542,18 +542,22 @@ function UnInstall-OpenSSH
         [string]$OpenSSHDir = "$env:SystemDrive\OpenSSH"
     )
 
+    if (-not (Test-Path $OpenSSHDir))
+    {
+        return
+    }
+
     Push-Location $OpenSSHDir
     if((Get-Service ssh-agent -ErrorAction Ignore) -ne $null) {
         Stop-Service ssh-agent -Force
     }
     &( "$OpenSSHDir\uninstall-sshd.ps1")
-    &( "$OpenSSHDir\uninstall-sshlsa.ps1")
         
     $machinePath = [Environment]::GetEnvironmentVariable('Path', 'MACHINE')
     $newMachineEnvironmentPath = $machinePath
     if ($machinePath.ToLower().Contains($OpenSSHDir.ToLower()))
     {        
-        $newMachineEnvironmentPath.Replace("$OpenSSHDir;", '')
+        $newMachineEnvironmentPath = $newMachineEnvironmentPath.Replace("$OpenSSHDir;", '')
         $env:Path = $env:Path.Replace("$OpenSSHDir;", '')
     }
